Make all operations atomic
This makes all operations atomic. Previously, this was done in a sort of
piecemeal fashion. This just cleans this up and makes sure all
operations are wrapped in the generic diesel transaction. This is
available for all Diesel backends that implement Diesel::Connection,
including the currently supported sqlite and postgres backends.

Signed-off-by: Davey Newhall <newhall@bitwise.io>